Move save&restore feature to new file usr.sbin/bhyve/migration
