+++ usr.sbin/bhyve/bhyverun.h
@@ -36,19 +36,6 @@
 #define	VMEXIT_CONTINUE		(0)
 #define	VMEXIT_ABORT		(-1)
 
-struct __attribute__((packed)) restore_state {
-	int kdata_fd;
-	int vmmem_fd;
-
-	void *kdata_map;
-	size_t kdata_len;
-
-	size_t vmmem_len;
-
-	struct ucl_parser *meta_parser;
-	ucl_object_t *meta_root_obj;
-};
-
 struct vmctx;
 extern int guest_ncpus;
 extern char *guest_uuid_str;
